LayoutTests/imported/w3c/web-platform-tests/css/motion/animation/offset-path-coord-box-interpolation-expected.txt


FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (-0.1) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from - 0.1px - 0.1px , line to 9px 9px ) "
PASS CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0) should be [shape(from 0px 0px, line to 10px 10px)]
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0.4) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0.4px 0.4px , line to 14px 14px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0.5) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 0.5px 0.5px , line to 15px 15px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (1) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 1px 1px , line to 20px 20px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (1.5) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 1.5px 1.5px , line to 25px 25px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (-0.1) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from - 0.1px - 0.1px , line to 9px 9px ) "
PASS CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0) should be [shape(from 0px 0px, line to 10px 10px)]
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0.4) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0.4px 0.4px , line to 14px 14px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0.5) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 0.5px 0.5px , line to 15px 15px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (1) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 1px 1px , line to 20px 20px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (1.5) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 1.5px 1.5px , line to 25px 25px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (-0.1) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from - 0.1px - 0.1px , line to 9px 9px ) "
PASS CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0) should be [shape(from 0px 0px, line to 10px 10px)]
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0.4) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0.4px 0.4px , line to 14px 14px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0.5) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 0.5px 0.5px , line to 15px 15px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (1) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 1px 1px , line to 20px 20px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (1.5) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 1.5px 1.5px , line to 25px 25px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (-0.1) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from - 0.1px - 0.1px , line to 9px 9px ) "
PASS Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0) should be [shape(from 0px 0px, line to 10px 10px)]
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0.4) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0.4px 0.4px , line to 14px 14px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (0.5) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 0.5px 0.5px , line to 15px 15px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (1) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 1px 1px , line to 20px 20px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) border-box] to [shape(from 1px 1px, line to 20px 20px) content-box] at (1.5) should be [shape(from 1px 1px, line to 20px 20px) content-box] assert_equals: expected "shape ( from 1px 1px , line to 20px 20px ) content - box " but got "shape ( from 1.5px 1.5px , line to 25px 25px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (-0.1) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0px 0px , line to 9px 9px ) "
PASS CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0) should be [shape(from 0px 0px, line to 10px 10px)]
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0.4) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0px 0px , line to 14px 14px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0.5) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 15px 15px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (1) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 20px 20px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (1.5) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 25px 25px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (-0.1) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0px 0px , line to 9px 9px ) "
PASS CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0) should be [shape(from 0px 0px, line to 10px 10px)]
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0.4) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0px 0px , line to 14px 14px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0.5) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 15px 15px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (1) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 20px 20px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (1.5) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 25px 25px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (-0.1) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0px 0px , line to 9px 9px ) "
PASS CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0) should be [shape(from 0px 0px, line to 10px 10px)]
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0.4) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0px 0px , line to 14px 14px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0.5) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 15px 15px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (1) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 20px 20px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (1.5) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 25px 25px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (-0.1) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0px 0px , line to 9px 9px ) "
PASS Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0) should be [shape(from 0px 0px, line to 10px 10px)]
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0.4) should be [shape(from 0px 0px, line to 10px 10px)] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) " but got "shape ( from 0px 0px , line to 14px 14px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (0.5) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 15px 15px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (1) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 20px 20px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px)] to [shape(from 0px 0px, line to 20px 20px) padding-box] at (1.5) should be [shape(from 0px 0px, line to 20px 20px) padding-box] assert_equals: expected "shape ( from 0px 0px , line to 20px 20px ) padding - box " but got "shape ( from 0px 0px , line to 25px 25px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (-0.1) should be [shape(from -1px -1px, line to 9px 9px) view-box] assert_equals: expected "shape ( from - 1px - 1px , line to 9px 9px ) view - box " but got "shape ( from - 1px - 1px , line to 9px 9px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0) should be [shape(from 0px 0px, line to 10px 10px) view-box] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) view - box " but got "shape ( from 0px 0px , line to 10px 10px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0.4) should be [shape(from 4px 4px, line to 14px 14px) view-box] assert_equals: expected "shape ( from 4px 4px , line to 14px 14px ) view - box " but got "shape ( from 4px 4px , line to 14px 14px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0.5) should be [shape(from 5px 5px, line to 15px 15px) view-box] assert_equals: expected "shape ( from 5px 5px , line to 15px 15px ) view - box " but got "shape ( from 5px 5px , line to 15px 15px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (1) should be [shape(from 10px 10px, line to 20px 20px) view-box] assert_equals: expected "shape ( from 10px 10px , line to 20px 20px ) view - box " but got "shape ( from 10px 10px , line to 20px 20px ) "
FAIL CSS Transitions: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (1.5) should be [shape(from 15px 15px, line to 25px 25px) view-box] assert_equals: expected "shape ( from 15px 15px , line to 25px 25px ) view - box " but got "shape ( from 15px 15px , line to 25px 25px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (-0.1) should be [shape(from -1px -1px, line to 9px 9px) view-box] assert_equals: expected "shape ( from - 1px - 1px , line to 9px 9px ) view - box " but got "shape ( from - 1px - 1px , line to 9px 9px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0) should be [shape(from 0px 0px, line to 10px 10px) view-box] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) view - box " but got "shape ( from 0px 0px , line to 10px 10px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0.4) should be [shape(from 4px 4px, line to 14px 14px) view-box] assert_equals: expected "shape ( from 4px 4px , line to 14px 14px ) view - box " but got "shape ( from 4px 4px , line to 14px 14px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0.5) should be [shape(from 5px 5px, line to 15px 15px) view-box] assert_equals: expected "shape ( from 5px 5px , line to 15px 15px ) view - box " but got "shape ( from 5px 5px , line to 15px 15px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (1) should be [shape(from 10px 10px, line to 20px 20px) view-box] assert_equals: expected "shape ( from 10px 10px , line to 20px 20px ) view - box " but got "shape ( from 10px 10px , line to 20px 20px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (1.5) should be [shape(from 15px 15px, line to 25px 25px) view-box] assert_equals: expected "shape ( from 15px 15px , line to 25px 25px ) view - box " but got "shape ( from 15px 15px , line to 25px 25px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (-0.1) should be [shape(from -1px -1px, line to 9px 9px) view-box] assert_equals: expected "shape ( from - 1px - 1px , line to 9px 9px ) view - box " but got "shape ( from - 1px - 1px , line to 9px 9px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0) should be [shape(from 0px 0px, line to 10px 10px) view-box] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) view - box " but got "shape ( from 0px 0px , line to 10px 10px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0.4) should be [shape(from 4px 4px, line to 14px 14px) view-box] assert_equals: expected "shape ( from 4px 4px , line to 14px 14px ) view - box " but got "shape ( from 4px 4px , line to 14px 14px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0.5) should be [shape(from 5px 5px, line to 15px 15px) view-box] assert_equals: expected "shape ( from 5px 5px , line to 15px 15px ) view - box " but got "shape ( from 5px 5px , line to 15px 15px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (1) should be [shape(from 10px 10px, line to 20px 20px) view-box] assert_equals: expected "shape ( from 10px 10px , line to 20px 20px ) view - box " but got "shape ( from 10px 10px , line to 20px 20px ) "
FAIL CSS Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (1.5) should be [shape(from 15px 15px, line to 25px 25px) view-box] assert_equals: expected "shape ( from 15px 15px , line to 25px 25px ) view - box " but got "shape ( from 15px 15px , line to 25px 25px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (-0.1) should be [shape(from -1px -1px, line to 9px 9px) view-box] assert_equals: expected "shape ( from - 1px - 1px , line to 9px 9px ) view - box " but got "shape ( from - 1px - 1px , line to 9px 9px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0) should be [shape(from 0px 0px, line to 10px 10px) view-box] assert_equals: expected "shape ( from 0px 0px , line to 10px 10px ) view - box " but got "shape ( from 0px 0px , line to 10px 10px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0.4) should be [shape(from 4px 4px, line to 14px 14px) view-box] assert_equals: expected "shape ( from 4px 4px , line to 14px 14px ) view - box " but got "shape ( from 4px 4px , line to 14px 14px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (0.5) should be [shape(from 5px 5px, line to 15px 15px) view-box] assert_equals: expected "shape ( from 5px 5px , line to 15px 15px ) view - box " but got "shape ( from 5px 5px , line to 15px 15px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (1) should be [shape(from 10px 10px, line to 20px 20px) view-box] assert_equals: expected "shape ( from 10px 10px , line to 20px 20px ) view - box " but got "shape ( from 10px 10px , line to 20px 20px ) "
FAIL Web Animations: property <offset-path> from [shape(from 0px 0px, line to 10px 10px) view-box] to [shape(from 10px 10px, line to 20px 20px) view-box] at (1.5) should be [shape(from 15px 15px, line to 25px 25px) view-box] assert_equals: expected "shape ( from 15px 15px , line to 25px 25px ) view - box " but got "shape ( from 15px 15px , line to 25px 25px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (-0.1) should be [circle(10px)] assert_equals: expected "circle ( 10px ) " but got "circle ( 8px ) "
PASS CSS Transitions: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0) should be [circle(10px)]
FAIL CSS Transitions: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0.4) should be [circle(10px)] assert_equals: expected "circle ( 10px ) " but got "circle ( 18px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0.5) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 20px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (1) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 30px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (1.5) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 40px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (-0.1) should be [circle(10px)] assert_equals: expected "circle ( 10px ) " but got "circle ( 8px ) "
PASS CSS Transitions with transition: all: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0) should be [circle(10px)]
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0.4) should be [circle(10px)] assert_equals: expected "circle ( 10px ) " but got "circle ( 18px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0.5) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 20px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (1) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 30px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (1.5) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 40px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (-0.1) should be [circle(10px)] assert_equals: expected "circle ( 10px ) " but got "circle ( 8px ) "
PASS CSS Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0) should be [circle(10px)]
FAIL CSS Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0.4) should be [circle(10px)] assert_equals: expected "circle ( 10px ) " but got "circle ( 18px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0.5) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 20px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (1) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 30px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (1.5) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 40px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (-0.1) should be [circle(10px)] assert_equals: expected "circle ( 10px ) " but got "circle ( 8px ) "
PASS Web Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0) should be [circle(10px)]
FAIL Web Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0.4) should be [circle(10px)] assert_equals: expected "circle ( 10px ) " but got "circle ( 18px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (0.5) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 20px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (1) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 30px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) border-box] to [circle(30px) content-box] at (1.5) should be [circle(30px) content-box] assert_equals: expected "circle ( 30px ) content - box " but got "circle ( 40px ) "
PASS CSS Transitions: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (-0.1) should be [circle(20px)]
PASS CSS Transitions: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0) should be [circle(20px)]
PASS CSS Transitions: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0.4) should be [circle(20px)]
FAIL CSS Transitions: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0.5) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL CSS Transitions: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (1) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL CSS Transitions: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (1.5) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
PASS CSS Transitions with transition: all: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (-0.1) should be [circle(20px)]
PASS CSS Transitions with transition: all: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0) should be [circle(20px)]
PASS CSS Transitions with transition: all: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0.4) should be [circle(20px)]
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0.5) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (1) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (1.5) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
PASS CSS Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (-0.1) should be [circle(20px)]
PASS CSS Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0) should be [circle(20px)]
PASS CSS Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0.4) should be [circle(20px)]
FAIL CSS Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0.5) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL CSS Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (1) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL CSS Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (1.5) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
PASS Web Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (-0.1) should be [circle(20px)]
PASS Web Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0) should be [circle(20px)]
PASS Web Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0.4) should be [circle(20px)]
FAIL Web Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (0.5) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL Web Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (1) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL Web Animations: property <offset-path> from [circle(20px)] to [circle(20px) padding-box] at (1.5) should be [circle(20px) padding-box] assert_equals: expected "circle ( 20px ) padding - box " but got "circle ( 20px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (-0.1) should be [circle(8px) view-box] assert_equals: expected "circle ( 8px ) view - box " but got "circle ( 8px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0) should be [circle(10px) view-box] assert_equals: expected "circle ( 10px ) view - box " but got "circle ( 10px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0.4) should be [circle(18px) view-box] assert_equals: expected "circle ( 18px ) view - box " but got "circle ( 18px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0.5) should be [circle(20px) view-box] assert_equals: expected "circle ( 20px ) view - box " but got "circle ( 20px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (1) should be [circle(30px) view-box] assert_equals: expected "circle ( 30px ) view - box " but got "circle ( 30px ) "
FAIL CSS Transitions: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (1.5) should be [circle(40px) view-box] assert_equals: expected "circle ( 40px ) view - box " but got "circle ( 40px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (-0.1) should be [circle(8px) view-box] assert_equals: expected "circle ( 8px ) view - box " but got "circle ( 8px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0) should be [circle(10px) view-box] assert_equals: expected "circle ( 10px ) view - box " but got "circle ( 10px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0.4) should be [circle(18px) view-box] assert_equals: expected "circle ( 18px ) view - box " but got "circle ( 18px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0.5) should be [circle(20px) view-box] assert_equals: expected "circle ( 20px ) view - box " but got "circle ( 20px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (1) should be [circle(30px) view-box] assert_equals: expected "circle ( 30px ) view - box " but got "circle ( 30px ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (1.5) should be [circle(40px) view-box] assert_equals: expected "circle ( 40px ) view - box " but got "circle ( 40px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (-0.1) should be [circle(8px) view-box] assert_equals: expected "circle ( 8px ) view - box " but got "circle ( 8px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0) should be [circle(10px) view-box] assert_equals: expected "circle ( 10px ) view - box " but got "circle ( 10px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0.4) should be [circle(18px) view-box] assert_equals: expected "circle ( 18px ) view - box " but got "circle ( 18px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0.5) should be [circle(20px) view-box] assert_equals: expected "circle ( 20px ) view - box " but got "circle ( 20px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (1) should be [circle(30px) view-box] assert_equals: expected "circle ( 30px ) view - box " but got "circle ( 30px ) "
FAIL CSS Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (1.5) should be [circle(40px) view-box] assert_equals: expected "circle ( 40px ) view - box " but got "circle ( 40px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (-0.1) should be [circle(8px) view-box] assert_equals: expected "circle ( 8px ) view - box " but got "circle ( 8px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0) should be [circle(10px) view-box] assert_equals: expected "circle ( 10px ) view - box " but got "circle ( 10px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0.4) should be [circle(18px) view-box] assert_equals: expected "circle ( 18px ) view - box " but got "circle ( 18px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (0.5) should be [circle(20px) view-box] assert_equals: expected "circle ( 20px ) view - box " but got "circle ( 20px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (1) should be [circle(30px) view-box] assert_equals: expected "circle ( 30px ) view - box " but got "circle ( 30px ) "
FAIL Web Animations: property <offset-path> from [circle(10px) view-box] to [circle(30px) view-box] at (1.5) should be [circle(40px) view-box] assert_equals: expected "circle ( 40px ) view - box " but got "circle ( 40px ) "
